Add RESERVED_BATCH_KEYS
torchtune passes a dictionary to its collate functions, for example padded_collate accepts a batch of type List[Dict[str, List[int]]] and returns a dictionary of tensors. In order to return the IDs of samples in a batch, we need to return this from the dataloader to relate model generations back to GT outputs e.g. ground truth transcripts for ASR to compute WER. Do this by compiling a dictionary of the expected key-value pairs (for the collate function and downstream dataloader) and the original items (e.g. ID field) but make it safe by checking that we do not have overlapping keys.