Demo: Show maximum sequence lengths for each iteration
Show the max. seq. lengths for each batch (iteration) i.e. within steps, given we are using gradient accumulation.

This demo is here as the online interleaving caused the training loop to go out of memory (OOM) which was not the case with the offline-computed MLS 25% en-xx stratified sample of the train split. This demo shows that maximum sequence lengths within a batch are typically in the range 400-500 with many in 300-400. Very few or none were >800 when eyeballing, which they are for the online-interleaved DSU-text MLS