Need to use a longer default sequence length for the ASR-style data
The ASR-style "instruction tuning" data consists of systems prompt, DSUs and text tokens (transcript), plus any eventual prompt template we end up using

Note: In testing, this was found to be fine for all the samples I checked, i.e. there was no truncation of the sample before the text and indeed the EOS token appeared; usually I identified this according to its token ID: 13303 in _these_ checks. I also saw a True in the mask at the final position after several False entries; since the Llama3Tokenizer tokenize_messages method truncates the mask to the same length as the tokens field, this is also a valid means of identifying that the sequence has not been prematurely truncated