Refactor out the llama_config resolution from model setup
Motivation: Separate the update of the "Llam Config" object which contains the model hyperparameters from the model setup. Note: We use a singleton for the Llama config - the configllama3_2_1b object - which is passed around over the lifetime of a single Python process. (Singleton implementation is based on Python import system which will mean that the object is instantiated only once when the module is loaded. Hopefully this is correct!)

Motivation: I needed to use the config object in the vLLM-based generation script and obviously did not want to instantiate a redundant torchtune model (on top of the one vLLM will create for me based on my HF-compatible files). The entangling of the config update and model setup was also muddy.

Refactored code allows modular use of the Llama config and clearly communicates intent + shows the update of the model hyperparameters at nearly "top-level" fucntions like `train` or `generate`.