Fix: Remove buggy torchtune wandb_logger.log_config call
This function "logs" the config to the checkpointer.checkpoint_dir directory, which in our case is a common directory used for all CPT/SFT runs that start from the published Llama 3.2 1B model - we only want/need to keep one copy of this base model around. The bug arises from the fact that subsequent training runs (experiments) will also write their configs to this common directory. When checkpoints are saved, all non-excluded files are copied from the checkpointer.checkpoint_dir to the checkpoint output directory (where weights for a given step are saved), including whichever torchtune_config.yaml is present in the common directory at the moment of checkpointing.

We observe that a single run can have multiple different values for the dataset, LR, warmup steps, and any other parameter that is changed in a subsequent training run

In testing the removal of this line, we see that W&B still receives a the parameters used for a training run, therefore we can rely on the configuration present in W&B to infer parameters for inference (for now)